requirements.txt
discord.py==2.3.2
psutil
greenlet
ipykernel==6.29.2
ipython==8.21.0
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
scipy==1.12.0